Fixed wrong name for barebones theme and renamed barebones theme file
